+++ tests/data/five11/operators.py
@@ -1,0 +1,17 @@
+DATA = [
+    {
+        "Id": "SC",
+        "Name": "Santa Clara Valley Transportation Authority",
+        "Monitored": True,
+    },
+    {
+        "Id": "SF",
+        "Name": "San Francisco Municipal Transportation Agency",
+        "Monitored": True,
+    },
+    {
+        "Id": "CT",
+        "Name": "Caltrain",
+        "Monitored": True,
+    },
+]