+++ tests/data/five11/stop_monitoring.py
@@ -1,0 +1,48 @@
+DATA = {
+    "ServiceDelivery": {
+        "ResponseTimestamp": "2026-07-03T12:00:00Z",
+        "ProducerRef": "511",
+        "Status": True,
+        "StopMonitoringDelivery": {
+            "version": "1.4",
+            "ResponseTimestamp": "2026-07-03T12:00:00Z",
+            "Status": True,
+            "MonitoredStopVisit": [
+                {
+                    "RecordedAtTime": "2026-07-03T12:00:00Z",
+                    "MonitoringRef": "70021",
+                    "MonitoredVehicleJourney": {
+                        "LineRef": "22",
+                        "DirectionRef": "N",
+                        "PublishedLineName": "22",
+                        "OperatorRef": "SC",
+                        "DestinationName": "Palo Alto Transit Center",
+                        "MonitoredCall": {
+                            "StopPointRef": "70021",
+                            "StopPointName": "El Camino Real & Palo Alto Ave",
+                            "AimedArrivalTime": "2099-01-01T00:04:00Z",
+                            "ExpectedArrivalTime": "2099-01-01T00:05:00Z"
+                        }
+                    }
+                },
+                {
+                    "RecordedAtTime": "2026-07-03T12:00:00Z",
+                    "MonitoringRef": "70021",
+                    "MonitoredVehicleJourney": {
+                        "LineRef": "522",
+                        "DirectionRef": "N",
+                        "PublishedLineName": "522 Rapid",
+                        "OperatorRef": "SC",
+                        "DestinationName": "Palo Alto Transit Center",
+                        "MonitoredCall": {
+                            "StopPointRef": "70021",
+                            "StopPointName": "El Camino Real & Palo Alto Ave",
+                            "AimedArrivalTime": "2099-01-01T00:11:00Z",
+                            "ExpectedArrivalTime": "2099-01-01T00:12:00Z"
+                        }
+                    }
+                }
+            ]
+        }
+    }
+}